+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+## 1.1.1 / 2016-09-05
+
+* [BUGFIX]  Don't fail when HTML contains valid unclosed tags eg <input>
+  by parsing as HTML rather than XML. This unfortunately does mean the 
+  HTML parser is a lot more tolerant than it was and will usually not
+  detect invalid HTML markup within the tables.
+
 ## 1.1.0 / 2015-10-15
 
 * [FEATURE] Parse multiple colspan columns in HTML tables as `...`